% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{get_SE}
\alias{get_SE}
\title{Parametric bootstrap standard error estimation}
\usage{
get_SE(x, ...)
}
\arguments{
\item{x}{an object for which a standard error estimate is desired}

\item{...}{other arguments.}
}
\value{
A list with the following components:
\item{est.se}{Estimated standard error of the mean estimator.}
\item{boot.means}{Bootstrap replicates of the mean estimates.}
\item{boot.sds}{Bootstrap replicates of the standard deviation estimates.}
}
\description{
A generic function for computing parametric bootstrap standard error (SE) estimates.
}
\references{
McGrath S., Katzenschlager S., Zimmer A.J., Seitel A., Steele R., Benedetti A. (In preparation). Standard error estimation in meta-analysis of studies reporting medians.
}
